Update truncate-plugin-titles.js
Prompt -  Since the image for the plugin is in the same div as the plugin title when you are truncating it you are also removing the thumbnail image, can you exclude img tags from the truncation?

ChatGTP -
My apologies for the oversight. To ensure the thumbnail image is not affected during the truncation process, we can update the truncate-plugin-titles.js file to only modify the text content of the plugin title without altering any child elements.

Here's the updated JavaScript code for the truncate-plugin-titles.js file: