Update truncate-plugin-titles.js
Prompt - it looks like the title is being truncated but the title isnt a link any more and the image keeps getting removed, can you fix it?

I apologize for the inconvenience. Here's an updated version of the code that should truncate the title without breaking the link or removing the image.

Please replace the content of the truncate-plugin-titles.js file with the following code: